Itroduced the DataProgram to be inherited. This loads data sets as required.
